test(daemon): give bgOutputFixture its own file
Appending it to testhelpers_test.go dropped that file's cohesion 8.55 -> 6.90
(CodeScene: Low Cohesion on a hotspot) — it is a self-contained fixture, not
another entry in the shared-mock grab-bag. Own file restores testhelpers_test.go
to its previous shape and gives the fixture a cohesive home.
